move pool.go under app/pub package
